requirements.txt
numpy
pandas
streamlit
streamlit_option_menu
streamlit_extras
PyPDF2
google-generativeai
openpyxl
serpapi
